requirements.txt
astroid==2.11.7
click==8.1.3
colorama==0.4.5
dill==0.3.5.1
Flask==2.1.3
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.7.0
numpy==1.23.1
pandas==1.4.3
platformdirs==2.5.2
plotly==5.9.0
pylint==2.14.5
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
six==1.16.0
SQLAlchemy==1.4.39
tenacity==8.0.1
tomli==2.0.1
tomlkit==0.11.1
Werkzeug==2.2.1
wrapt==1.14.1
